requirements.txt
configargparse
cycler
einops
matplotlib
mediapy
numba
numpy==1.26.4
open3d
Pillow
plyfile
pyrender
scikit_learn
scipy
setuptools==59.5.0
scikit-image
tqdm
trimesh
tensorboard
opencv-python==4.10.0.84
